backend/requirements.txt
### Django에 필요한 의존 패키지들

# 사용할 Django 버전
Django==4.2.11

# postgreSQL과 통신을 위해 필요한 패키지(psycopg3)
psycopg2-binary

# channel
channels
# daphne
daphne

djangorestframework
djangorestframework-simplejwt
<<<<<<< HEAD
requests
channels_postgres
=======
requests
>>>>>>> 390bbfae75df8f255d8512756a932f5fad670586
